Update to python 3.10 and update package versions (#1)
* Python 3.10 and suggested updates from Ray Jones; install mamba

* Bump to version 0.0.2

* Update google-cloud-cli version

* Install compiled crcmod

* Update package versions and install libarchive for mamba

+++ build.sh
@@ -4,6 +4,7 @@
 set -e -x
 
 IMAGE_DIR=$1
+SKIP_GCP_AUTH=$2
 VERSION=$(cat config/conf.json | jq -r ".image_data | .[] | select(.name == \"$IMAGE_DIR\") | .version")
 
 TAG_NAME=$(git log --pretty=format:'%h' -n 1)
@@ -25,13 +26,16 @@ if [[ $VAULT_LOCATION == *"jenkins"* ]]; then
     VAULT_LOCATION="/etc/vault-token-dsde"
 fi
 
-# will fail if you are not gcloud authed as dspci-wb-gcr-service-account or secret/dsde/firecloud/common/image-build-account.json
-# the below works locally but not on jenkins for some reason, using a different account until a resolution is found
-# docker run --rm  -v $VAULT_LOCATION:/root/.vault-token:ro broadinstitute/dsde-toolbox:latest vault read --format=json secret/dsde/dsp-techops/common/dspci-wb-gcr-service-account.json | jq .data > dspci-wb-gcr-service-account.json
-# gcloud auth activate-service-account --key-file=dspci-wb-gcr-service-account.json
-docker run --rm  -v $VAULT_LOCATION:/root/.vault-token:ro broadinstitute/dsde-toolbox:latest vault read --format=json secret/dsde/firecloud/common/image-build-account.json | jq .data > image-build-account.json
-gcloud auth activate-service-account --key-file=image-build-account.json
-gcloud auth configure-docker --quiet
+if [[ -z "${SKIP_GCP_AUTH}" ]] || [["${SKIP_GCP_AUTH}" != "true"]]; then
+  # will fail if you are not gcloud authed as dspci-wb-gcr-service-account or secret/dsde/firecloud/common/image-build-account.json
+  # the below works locally but not on jenkins for some reason, using a different account until a resolution is found
+  # docker run --rm  -v $VAULT_LOCATION:/root/.vault-token:ro broadinstitute/dsde-toolbox:latest vault read --format=json secret/dsde/dsp-techops/common/dspci-wb-gcr-service-account.json | jq .data > dspci-wb-gcr-service-account.json
+  # gcloud auth activate-service-account --key-file=dspci-wb-gcr-service-account.json
+  docker run --rm  -v $VAULT_LOCATION:/root/.vault-token:ro broadinstitute/dsde-toolbox:latest vault read --format=json secret/dsde/firecloud/common/image-build-account.json | jq .data > image-build-account.json
+  gcloud auth activate-service-account --key-file=image-build-account.json
+  gcloud auth configure-docker --quiet
+fi
+
 
 docker image build ./$IMAGE_DIR \
   --tag $GCR_IMAGE_REPO/$IMAGE_DIR:$TAG_NAME \
@@ -42,22 +46,14 @@ docker push $GCR_IMAGE_REPO/$IMAGE_DIR:$TAG_NAME
 docker push $GCR_IMAGE_REPO/$IMAGE_DIR:$VERSION
 docker push $GCR_IMAGE_REPO/$IMAGE_DIR:latest
 
-if [[ $IMAGE_DIR = "terra-jupyter-aou" ]]; then
-  docker tag $GCR_IMAGE_REPO/$IMAGE_DIR:$TAG_NAME broadinstitute/$IMAGE_DIR:$TAG_NAME
-  docker tag $GCR_IMAGE_REPO/$IMAGE_DIR:$VERSION broadinstitute/$IMAGE_DIR:$VERSION
-  docker tag $GCR_IMAGE_REPO/$IMAGE_DIR:$VERSION broadinstitute/$IMAGE_DIR:latest
-
-  docker push broadinstitute/$IMAGE_DIR:$VERSION
-  docker push broadinstitute/$IMAGE_DIR:$TAG_NAME
-  docker push broadinstitute/$IMAGE_DIR:latest
-fi
-
 docker kill $IMAGE_DIR || true
 docker rm -f $IMAGE_DIR || true
 docker run --rm -itd -u root -e PIP_USER=false --entrypoint='/bin/bash' --name $IMAGE_DIR $GCR_IMAGE_REPO/$IMAGE_DIR:$VERSION
 
-gcloud auth list
-python scripts/generate_package_docs.py "$IMAGE_DIR"
+if [[ $IMAGE_DIR != "terra-rstudio-aou" ]]; then
+  gcloud auth list
+  python scripts/generate_package_docs.py "$IMAGE_DIR"
+fi
 
 docker kill $IMAGE_DIR || true
 docker rm -f $IMAGE_DIR || true
